+++ quic.cabal
@@ -53,6 +53,7 @@ library
                        Network.QUIC.Connection.Types
                        Network.QUIC.Connector
                        Network.QUIC.Crypto
+                       Network.QUIC.Crypto.Types
                        Network.QUIC.CryptoFusion
                        Network.QUIC.Exception
                        Network.QUIC.Handshake
